[chore] Improve documentation and split clients
