+++ redpanda/resources/topic/topic_resource.go
@@ -159,8 +159,10 @@ func resourceTopicSchema() schema.Schema {
 				Required: true,
 			},
 			"cluster_api_url": schema.StringAttribute{
-				Required:      true,
-				Description:   "The cluster API URL",
+				Required: true,
+				Description: "The cluster API URL. Changing this will prevent deletion of the resource on the existing " +
+					"cluster. It is generally a better idea to delete an existing resource and create a new one than to " +
+					"change this value unless you are planning to do state imports",
 				PlanModifiers: []planmodifier.String{stringplanmodifier.RequiresReplace()},
 			},
 		},
